+++ modules/capacity.py
@@ -45,16 +45,16 @@ def print_shape_list(home_region, config, signer, compartment_id):
     This list is automatically refreshed when a new shape is detected in the connected region.
     """
 
-    all_shapes = [
-    'BM.DenseIO1.36', 'BM.DenseIO2.52', 'BM.DenseIO.E4.128', 'BM.DenseIO.E5.128', 'BM.GPU2.2', 'BM.GPU3.8', 'BM.GPU4.8', 'BM.GPU.A10.4', 
-    'BM.GPU.A100-v2.8', 'BM.GPU.H100.8', 'BM.GPU.L40S.4', 'BM.HPC2.36', 'BM.HPC.E5.144', 'BM.Optimized3.36', 'BM.Standard1.36', 'BM.Standard2.52', 
-    'BM.Standard3.64', 'BM.Standard.A1.160', 'BM.Standard.B1.44', 'BM.Standard.E2.64', 'BM.Standard.E3.128', 'BM.Standard.E4.128', 'BM.Standard.E5.192', 
-    'VM.DenseIO1.16', 'VM.DenseIO1.4', 'VM.DenseIO1.8', 'VM.DenseIO2.16', 'VM.DenseIO2.24', 'VM.DenseIO2.8', 'VM.DenseIO.E4.Flex', 'VM.DenseIO.E5.Flex', 
-    'VM.GPU2.1', 'VM.GPU3.1', 'VM.GPU3.2', 'VM.GPU3.4', 'VM.GPU.A10.1', 'VM.GPU.A10.2', 'VM.Optimized3.Flex', 'VM.Standard1.1', 'VM.Standard1.16', 
-    'VM.Standard1.2', 'VM.Standard1.4', 'VM.Standard1.8', 'VM.Standard2.1', 'VM.Standard2.16', 'VM.Standard2.2', 'VM.Standard2.24', 'VM.Standard2.4', 
-    'VM.Standard2.8', 'VM.Standard3.Flex', 'VM.Standard.A1.Flex', 'VM.Standard.A2.Flex', 'VM.Standard.B1.1', 'VM.Standard.B1.16', 'VM.Standard.B1.2', 
-    'VM.Standard.B1.4', 'VM.Standard.B1.8', 'VM.Standard.E2.1', 'VM.Standard.E2.1.Micro', 'VM.Standard.E2.2', 'VM.Standard.E2.4', 'VM.Standard.E2.8', 
-    'VM.Standard.E3.Flex', 'VM.Standard.E4.Flex', 'VM.Standard.E5.Flex'
+    known_shapes = [
+    'BM.DenseIO1.36', 'BM.DenseIO2.52', 'BM.DenseIO.E4.128', 'BM.DenseIO.E5.128', 'BM.GPU2.2', 'BM.GPU3.8', 'BM.GPU4.8', 'BM.GPU.A10.4', 'BM.GPU.A100-v2.8', 
+    'BM.GPU.H100.8', 'BM.GPU.L40S.4', 'BM.GPU.MI300X.8', 'BM.GPU.MI355X.8', 'BM.GPU.H200.8', 'BM.GPU.B200.8', 'BM.GPU.GB200.4', 'BM.GPU.GB300.4', 'BM.HPC2.36', 
+    'BM.HPC.E5.144', 'BM.Optimized3.36', 'BM.Standard1.36', 'BM.Standard2.52', 'BM.Standard3.64', 'BM.Standard.A1.160', 'BM.Standard.A4.48', 'BM.Standard.B1.44', 
+    'BM.Standard.E2.64', 'BM.Standard.E3.128', 'BM.Standard.E4.128', 'BM.Standard.E5.192', 'BM.Standard.E6.256', 'VM.DenseIO1.16', 'VM.DenseIO1.4', 'VM.DenseIO1.8', 
+    'VM.DenseIO2.16', 'VM.DenseIO2.24', 'VM.DenseIO2.8', 'VM.DenseIO.E4.Flex', 'VM.DenseIO.E5.Flex', 'VM.GPU2.1', 'VM.GPU3.1', 'VM.GPU3.2', 'VM.GPU3.4', 'VM.GPU.A10.1',
+    'VM.GPU.A10.2', 'VM.Optimized3.Flex', 'VM.Standard1.1', 'VM.Standard1.16', 'VM.Standard1.2', 'VM.Standard1.4', 'VM.Standard1.8', 'VM.Standard2.1', 'VM.Standard2.16', 
+    'VM.Standard2.2', 'VM.Standard2.24', 'VM.Standard2.4', 'VM.Standard2.8', 'VM.Standard3.Flex', 'VM.Standard.A1.Flex', 'VM.Standard.A2.Flex', 'VM.Standard.A4.Flex', 
+    'VM.Standard.B1.1', 'VM.Standard.B1.16', 'VM.Standard.B1.2', 'VM.Standard.B1.4', 'VM.Standard.B1.8', 'VM.Standard.E2.1', 'VM.Standard.E2.1.Micro', 'VM.Standard.E2.2', 
+    'VM.Standard.E2.4', 'VM.Standard.E2.8', 'VM.Standard.E3.Flex', 'VM.Standard.E4.Flex', 'VM.Standard.E5.Flex', 'VM.Standard.E6.Flex'
     ]
 
     try:
@@ -65,12 +65,12 @@ def print_shape_list(home_region, config, signer, compartment_id):
         # Fetch and sort available shapes in the region
         shapes_in_home_region = core_client.list_shapes(compartment_id).data
 
-        # Update all_shapes list if new shapes found
+        # Update known_shapes list if new shapes found
         for shape in shapes_in_home_region:
-            if shape.shape not in all_shapes:
-                all_shapes.append(shape.shape)
+            if shape.shape not in known_shapes:
+                known_shapes.append(shape.shape)
         
-        all_shapes = sorted(all_shapes)
+        all_shapes = sorted(known_shapes)
 
         print(yellow("\nGet all available shapes at: https://docs.oracle.com/en-us/iaas/Content/Compute/References/computeshapes.htm\n"))
 
@@ -197,7 +197,7 @@ def get_shape_config(user_shape_name, shapes_in_region, user_shape_ocpus, user_s
         return shape_ocpus, shape_memory, shape_is_flex, shape_info
     
     # A2.Flex: Shape-Specific configuration
-    elif user_shape_name == "VM.Standard.A2.Flex":
+    elif user_shape_name == "VM.Standard.A2.Flex" or user_shape_name == "VM.Standard.A4.Flex":
         shape_info = next((shape for shape in shapes_in_region if shape.shape == user_shape_name), None)
 
         if shape_info: